First pass at removing worktree creation logic out of the agent panel
