+++ crates/git_ui/src/worktree_names.rs
